Fix alarm leak on exceptional case
Because Go and C have some difference on timer, there are special case
on 'DoTasks()' fuction. While DoTask() function, Go recognizes the
alarm should ring but C does not. This case was described and handled
on 'alarm.go:44'. However, when root alarm is faced with this
situation, TaskRunner cannot reset its timer on resetTimer() function
and suffers problem with memory.